# Make cirq.Zip twenty times faster