--- a/cirq-core/cirq/study/sweeps_test.py
+++ b/cirq-core/cirq/study/sweeps_test.py
@@ -76,6 +76,11 @@ def test_zip():
     assert len(sweep) == 3
     assert _values(sweep, 'a') == [1, 2, 3]
     assert _values(sweep, 'b') == [4, 5, 6]
+    assert list(sweep.param_tuples()) == [
+        (('a', 1), ('b', 4)),
+        (('a', 2), ('b', 5)),
+        (('a', 3), ('b', 6)),
+    ]
 
 
 def test_zip_longest():