- cirq.Zip param_tuples was doing a sum of tuples, which was creating n^2 tuples over the course of the calculations.
- Use itertools to do this correctly.

Using a zip of 1000 different values of 10000 different values: Before: 110 seconds.  After: 5.6 seconds

```
%timeit sum(values, ())
111 ms ± 3.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit tuple(itertools.chain.from_iterable(values))
431 µs ± 3.98 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```