@maffoo 2025-04-29
Looks like `itertools.chain` wins, so let's stick with what you have written:
```python
In [1]: values = [tuple(range(10000)) for _ in range(1000)]

In [2]: import itertools

In [3]: %timeit tuple(itertools.chain.from_iterable(values))
130 ms ± 1.76 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [4]: %timeit tuple(param for params in values for param in params)
364 ms ± 5.89 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```